move localize name to common/helpers
